refactor: traverse flat_map
