+++ crates/nu-lsp/src/ast.rs
@@ -1,171 +1,10 @@
 use crate::Id;
 use nu_protocol::{
-    ast::{
-        Argument, Block, Call, Expr, Expression, ExternalArgument, ListItem, MatchPattern,
-        PathMember, Pattern, PipelineRedirection, RecordItem,
-    },
+    ast::{Argument, Block, Call, Expr, Expression, ListItem, PathMember, Traverse},
     engine::StateWorkingSet,
     Span,
 };
 use std::sync::Arc;
-
-/// similar to flatten_block, but allows extra map function
-pub fn ast_flat_map<'a, T, F>(
-    ast: &'a Arc<Block>,
-    working_set: &'a StateWorkingSet,
-    f_special: &F,
-) -> Vec<T>
-where
-    F: Fn(&'a Expression) -> Option<Vec<T>>,
-{
-    ast.pipelines
-        .iter()
-        .flat_map(|pipeline| {
-            pipeline.elements.iter().flat_map(|element| {
-                expr_flat_map(&element.expr, working_set, f_special)
-                    .into_iter()
-                    .chain(
-                        element
-                            .redirection
-                            .as_ref()
-                            .map(|redir| redirect_flat_map(redir, working_set, f_special))
-                            .unwrap_or_default(),
-                    )
-            })
-        })
-        .collect()
-}
-
-/// generic function that do flat_map on an expression
-/// concats all recursive results on sub-expressions
-///
-/// # Arguments
-/// * `f_special` - function that overrides the default behavior
-pub fn expr_flat_map<'a, T, F>(
-    expr: &'a Expression,
-    working_set: &'a StateWorkingSet,
-    f_special: &F,
-) -> Vec<T>
-where
-    F: Fn(&'a Expression) -> Option<Vec<T>>,
-{
-    // behavior overridden by f_special
-    if let Some(vec) = f_special(expr) {
-        return vec;
-    }
-    let recur = |expr| expr_flat_map(expr, working_set, f_special);
-    match &expr.expr {
-        Expr::RowCondition(block_id)
-        | Expr::Subexpression(block_id)
-        | Expr::Block(block_id)
-        | Expr::Closure(block_id) => {
-            let block = working_set.get_block(block_id.to_owned());
-            ast_flat_map(block, working_set, f_special)
-        }
-        Expr::Range(range) => [&range.from, &range.next, &range.to]
-            .iter()
-            .filter_map(|e| e.as_ref())
-            .flat_map(recur)
-            .collect(),
-        Expr::Call(call) => call
-            .arguments
-            .iter()
-            .filter_map(|arg| arg.expr())
-            .flat_map(recur)
-            .collect(),
-        Expr::ExternalCall(head, args) => recur(head)
-            .into_iter()
-            .chain(args.iter().flat_map(|arg| match arg {
-                ExternalArgument::Regular(e) | ExternalArgument::Spread(e) => recur(e),
-            }))
-            .collect(),
-        Expr::UnaryNot(expr) | Expr::Collect(_, expr) => recur(expr),
-        Expr::BinaryOp(lhs, op, rhs) => recur(lhs)
-            .into_iter()
-            .chain(recur(op))
-            .chain(recur(rhs))
-            .collect(),
-        Expr::MatchBlock(matches) => matches
-            .iter()
-            .flat_map(|(pattern, expr)| {
-                match_pattern_flat_map(pattern, working_set, f_special)
-                    .into_iter()
-                    .chain(recur(expr))
-            })
-            .collect(),
-        Expr::List(items) => items
-            .iter()
-            .flat_map(|item| match item {
-                ListItem::Item(expr) | ListItem::Spread(_, expr) => recur(expr),
-            })
-            .collect(),
-        Expr::Record(items) => items
-            .iter()
-            .flat_map(|item| match item {
-                RecordItem::Spread(_, expr) => recur(expr),
-                RecordItem::Pair(key, val) => [key, val].into_iter().flat_map(recur).collect(),
-            })
-            .collect(),
-        Expr::Table(table) => table
-            .columns
-            .iter()
-            .flat_map(recur)
-            .chain(table.rows.iter().flat_map(|row| row.iter().flat_map(recur)))
-            .collect(),
-        Expr::ValueWithUnit(vu) => recur(&vu.expr),
-        Expr::FullCellPath(fcp) => recur(&fcp.head),
-        Expr::Keyword(kw) => recur(&kw.expr),
-        Expr::StringInterpolation(vec) | Expr::GlobInterpolation(vec, _) => {
-            vec.iter().flat_map(recur).collect()
-        }
-
-        _ => Vec::new(),
-    }
-}
-
-/// flat_map on match patterns
-fn match_pattern_flat_map<'a, T, F>(
-    pattern: &'a MatchPattern,
-    working_set: &'a StateWorkingSet,
-    f_special: &F,
-) -> Vec<T>
-where
-    F: Fn(&'a Expression) -> Option<Vec<T>>,
-{
-    let recur = |expr| expr_flat_map(expr, working_set, f_special);
-    let recur_match = |p| match_pattern_flat_map(p, working_set, f_special);
-    match &pattern.pattern {
-        Pattern::Expression(expr) => recur(expr),
-        Pattern::List(patterns) | Pattern::Or(patterns) => {
-            patterns.iter().flat_map(recur_match).collect()
-        }
-        Pattern::Record(entries) => entries.iter().flat_map(|(_, p)| recur_match(p)).collect(),
-        _ => Vec::new(),
-    }
-    .into_iter()
-    .chain(pattern.guard.as_ref().map(|g| recur(g)).unwrap_or_default())
-    .collect()
-}
-
-/// flat_map on redirections
-fn redirect_flat_map<'a, T, F>(
-    redir: &'a PipelineRedirection,
-    working_set: &'a StateWorkingSet,
-    f_special: &F,
-) -> Vec<T>
-where
-    F: Fn(&'a Expression) -> Option<Vec<T>>,
-{
-    let recur = |expr| expr_flat_map(expr, working_set, f_special);
-    match redir {
-        PipelineRedirection::Single { target, .. } => target.expr().map(recur).unwrap_or_default(),
-        PipelineRedirection::Separate { out, err } => [out, err]
-            .iter()
-            .filter_map(|t| t.expr())
-            .flat_map(recur)
-            .collect(),
-    }
-}
 
 /// Adjust span if quoted
 fn strip_quotes(span: Span, working_set: &StateWorkingSet) -> Span {
@@ -512,7 +351,7 @@ pub(crate) fn find_id(
     location: &usize,
 ) -> Option<(Id, Span)> {
     let closure = |e| find_id_in_expr(e, working_set, location);
-    ast_flat_map(ast, working_set, &closure).first().cloned()
+    ast.flat_map(working_set, &closure).first().cloned()
 }
 
 fn find_reference_by_id_in_expr(
@@ -521,7 +360,6 @@ fn find_reference_by_id_in_expr(
     id: &Id,
 ) -> Option<Vec<Span>> {
     let closure = |e| find_reference_by_id_in_expr(e, working_set, id);
-    let recur = |expr| expr_flat_map(expr, working_set, &closure);
     match (&expr.expr, id) {
         (Expr::Var(vid1), Id::Variable(vid2)) if *vid1 == *vid2 => Some(vec![Span::new(
             // we want to exclude the `$` sign for renaming
@@ -536,7 +374,7 @@ fn find_reference_by_id_in_expr(
                 .arguments
                 .iter()
                 .filter_map(|arg| arg.expr())
-                .flat_map(recur)
+                .flat_map(|e| e.flat_map(working_set, &closure))
                 .collect();
             if matches!(id, Id::Declaration(decl_id) if call.decl_id == *decl_id) {
                 occurs.push(call.head);
@@ -560,7 +398,7 @@ pub(crate) fn find_reference_by_id(
     working_set: &StateWorkingSet,
     id: &Id,
 ) -> Vec<Span> {
-    ast_flat_map(ast, working_set, &|e| {
+    ast.flat_map(working_set, &|e| {
         find_reference_by_id_in_expr(e, working_set, id)
     })
 }
